add `Service` for handling groups of rpcs
first I only added a way to handle a single rpc request using a function
that recieves a message and sends a response. Now I want something
that can handle a bunch of RPCs (message types), run concurrently alongside the node
and keep its own state. I added a `Service` type which is a static
dispatch interface (don't quote me on that). but it is essentially an
interface for the functionality I described up there. A service is
anything that implements that interface.

Next is to figure out how to run this in concurrently using a separate
thread or something. I am getting a `cannot evaluate comptime value`
if I try to put `self.vtable.start` in a thread.